Made generalised container for dashboards and refractored appbar file

+++ lib/models/realtime_weather.dart
@@ -34,17 +34,17 @@ class Current {
   double windKph;
   int windDegree;
   String windDir;
-  int pressureMb;
+  double pressureMb;
   double pressureIn;
-  int precipMm;
-  int precipIn;
+  double precipMm;
+  double precipIn;
   int humidity;
   int cloud;
   double feelslikeC;
   double feelslikeF;
-  int visKm;
-  int visMiles;
-  int uv;
+  double visKm;
+  double visMiles;
+  double uv;
   double gustMph;
   double gustKph;
